Split runner project file write tests
